work/My projects/NovaIQ/backend/pyproject.toml
[project]
name = "rtc-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "sqlmodel>=0.0.22",
  "python-slugify>=8.0.4",
  "python-dateutil>=2.9.0.post0",
  "tenacity>=9.0.0",
  "orjson>=3.10.7",
]
# Chroma + OpenAI are optional; install when needed
[project.optional-dependencies]
vectors = ["chromadb>=0.5.5","openai>=1.51.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
addopts = "-q"
